+++ dweb-lib/src/web/name.rs
@@ -87,7 +87,7 @@ pub struct DwebHost {
     pub dweb_host_string: String,
     pub dweb_name: String,
     /// None implies most recent version (highest number)
-    pub version: Option<u32>,
+    pub version: Option<u64>,
 
     #[cfg(feature = "fixed-dweb-hosts")]
     // Development build feature for non-versioned Tree references
@@ -147,7 +147,7 @@ pub fn make_dweb_name(memorable_part: &String, history_address: HistoryAddress) 
 }
 
 /// Create a version part ("v[VERSION]") for a www-dweb URL
-pub fn make_version_part(version: u32) -> String {
+pub fn make_version_part(version: u64) -> String {
     if version > 0 {
         format!("v{version}")
     } else {
@@ -268,7 +268,7 @@ pub fn decode_dweb_host(dweb_host: &str) -> Result<DwebHost> {
                         "Dweb host contains four segments (separated by '.') so first must start with 'v'"
                     ));
                 }
-                match str[1..].parse::<u32>() {
+                match str[1..].parse::<u64>() {
                     Ok(version) => {
                         if version > 0 {
                             found_version_segment = true;
@@ -351,7 +351,7 @@ pub fn decode_dweb_host(dweb_host: &str) -> Result<DwebHost> {
 /// The part off a DWEB-NAME up but excluding the final hyphen is known as the 'memorable part'.
 ///
 /// The memorable_part must start with at least two alphabetic characters. This is to allow it to
-/// be distinguished from a version parameter, which is a 'v' or 'V' followed by an integer (u32),
+/// be distinguished from a version parameter, which is a 'v' or 'V' followed by an integer (u64),
 /// which is useful in apps, for parsing links where the version is an optional part of the URL path.
 ///
 /// Following the first two alphabetic characters are a number of alphanumeric characters which may
